win32: avoid warnings due to pid type
On win32, a process is identified by a HANDLE, but for printing
we should use GetProcessId() to return a meaningful integer.

Also call the handle to intptr_t for open_osfhandle() to avoid a compiler warning.

Fixes #217

+++ jimiocompat.c
@@ -121,7 +121,7 @@ int Jim_MakeTempFile(Jim_Interp *interp, const char *filename_template, int unli
     }
 
     Jim_SetResultString(interp, name, -1);
-    return _open_osfhandle((int)handle, _O_RDWR | _O_TEXT);
+    return _open_osfhandle((intptr_t)handle, _O_RDWR | _O_TEXT);
 
   error:
     Jim_SetResultErrno(interp, name);
